Add FacilityUserTquery class between AdminUserTquery and ClientTquery and StaffTquery to avoid injecting Facility which is not bound in the context of AdminUserTquery
