+++ app/Tquery/Tables/AdminUserTquery.php
@@ -2,9 +2,6 @@
 
 namespace App\Tquery\Tables;
 
-use App\Models\Facility;
-use App\Models\Meeting;
-use App\Models\MeetingAttendant;
 use App\Tquery\Config\TqConfig;
 use App\Tquery\Config\TqDataTypeEnum;
 use App\Tquery\Config\TqTableAliasEnum;
@@ -12,11 +9,6 @@ use App\Tquery\Engine\TqService;
 
 readonly class AdminUserTquery extends TqService
 {
-    public function __construct(protected Facility $facility)
-    {
-        parent::__construct();
-    }
-
     protected function getConfig(): TqConfig
     {
         $config = new TqConfig(table: TqTableAliasEnum::users);
@@ -51,118 +43,5 @@ readonly class AdminUserTquery extends TqService
         );
         $config->addCount();
         return $config;
-    }
-
-    protected function addMeetingsRelatedColumns(TqConfig $config): void
-    {
-        $c = fn($x) => '"' . $x . '"';
-
-        $config->addQuery(
-            TqDataTypeEnum::date_nullable,
-            fn(string $tableName) => //
-            "select min(`meetings`.`date`)
-                from `meetings`
-                inner join `meeting_attendants` on `meetings`.`id` = `meeting_attendants`.`meeting_id`
-                where `meetings`.`facility_id` = '{$this->facility->id}'
-                and `meeting_attendants`.`user_id` = `users`.`id`
-                and `meeting_attendants`.`attendance_status_dict_id` in
-                    ({$c(MeetingAttendant::ATTENDANCE_STATUS_OK)},
-                     {$c(MeetingAttendant::ATTENDANCE_STATUS_LATE_PRESENT)})
-                and `meetings`.`status_dict_id` = {$c(Meeting::STATUS_COMPLETED)}
-                and `meetings`.`category_dict_id` != {$c(Meeting::CATEGORY_SYSTEM)}
-                and `meetings`.`deleted_at` is null",
-            "first_meeting_date"
-        );
-
-        $config->addQuery(
-            TqDataTypeEnum::date_nullable,
-            fn(string $tableName) => //
-            "select max(`meetings`.`date`)
-                from `meetings`
-                inner join `meeting_attendants` on `meetings`.`id` = `meeting_attendants`.`meeting_id`
-                where `meetings`.`facility_id` = '{$this->facility->id}'
-                and `meeting_attendants`.`user_id` = `users`.`id`
-                and `meeting_attendants`.`attendance_status_dict_id` in
-                    ({$c(MeetingAttendant::ATTENDANCE_STATUS_OK)},
-                     {$c(MeetingAttendant::ATTENDANCE_STATUS_LATE_PRESENT)})
-                and `meetings`.`status_dict_id` = {$c(Meeting::STATUS_COMPLETED)}
-                and `meetings`.`category_dict_id` != {$c(Meeting::CATEGORY_SYSTEM)}
-                and `meetings`.`deleted_at` is null",
-            "last_meeting_date"
-        );
-
-        $config->addQuery(
-            TqDataTypeEnum::int,
-            fn(string $tableName) => //
-            "select count(1)
-                from `meetings`
-                inner join `meeting_attendants` on `meetings`.`id` = `meeting_attendants`.`meeting_id`
-                where `meetings`.`facility_id` = '{$this->facility->id}'
-                and `meeting_attendants`.`user_id` = `users`.`id`
-                and `meeting_attendants`.`attendance_status_dict_id` in
-                    ({$c(MeetingAttendant::ATTENDANCE_STATUS_OK)},
-                     {$c(MeetingAttendant::ATTENDANCE_STATUS_LATE_PRESENT)})
-                and `meetings`.`status_dict_id` = {$c(Meeting::STATUS_COMPLETED)}
-                and `meetings`.`category_dict_id` != {$c(Meeting::CATEGORY_SYSTEM)}
-                and `meetings`.`deleted_at` is null",
-            "completed_meetings_count"
-        );
-
-        $config->addQuery(
-            TqDataTypeEnum::int,
-            fn(string $tableName) => //
-            "select count(1)
-                from `meetings`
-                inner join `meeting_attendants` on `meetings`.`id` = `meeting_attendants`.`meeting_id`
-                where `meetings`.`facility_id` = '{$this->facility->id}'
-                and `meeting_attendants`.`user_id` = `users`.`id`
-                and `meeting_attendants`.`attendance_status_dict_id` in
-                    ({$c(MeetingAttendant::ATTENDANCE_STATUS_OK)},
-                     {$c(MeetingAttendant::ATTENDANCE_STATUS_LATE_PRESENT)})
-                and `meetings`.`status_dict_id` = {$c(Meeting::STATUS_COMPLETED)}
-                and `meetings`.`category_dict_id` != {$c(Meeting::CATEGORY_SYSTEM)}
-                and `meetings`.`deleted_at` is null
-                /* Last month */
-                and date > date_sub(curdate(), interval 1 month)
-                and date <= curdate()",
-            "completed_meetings_count_last_month"
-        );
-
-        $config->addQuery(
-            TqDataTypeEnum::int,
-            fn(string $tableName) => //
-            "select count(1)
-                from `meetings`
-                inner join `meeting_attendants` on `meetings`.`id` = `meeting_attendants`.`meeting_id`
-                where `meetings`.`facility_id` = '{$this->facility->id}'
-                and `meeting_attendants`.`user_id` = `users`.`id`
-                and `meeting_attendants`.`attendance_status_dict_id` in
-                    ({$c(MeetingAttendant::ATTENDANCE_STATUS_OK)},
-                     {$c(MeetingAttendant::ATTENDANCE_STATUS_LATE_PRESENT)})
-                and `meetings`.`status_dict_id` = {$c(Meeting::STATUS_PLANNED)}
-                and `meetings`.`category_dict_id` != {$c(Meeting::CATEGORY_SYSTEM)}
-                and `meetings`.`deleted_at` is null",
-            "planned_meetings_count"
-        );
-
-        $config->addQuery(
-            TqDataTypeEnum::int,
-            fn(string $tableName) => //
-            "select count(1)
-                from `meetings`
-                inner join `meeting_attendants` on `meetings`.`id` = `meeting_attendants`.`meeting_id`
-                where `meetings`.`facility_id` = '{$this->facility->id}'
-                and `meeting_attendants`.`user_id` = `users`.`id`
-                and `meeting_attendants`.`attendance_status_dict_id` in
-                    ({$c(MeetingAttendant::ATTENDANCE_STATUS_OK)},
-                     {$c(MeetingAttendant::ATTENDANCE_STATUS_LATE_PRESENT)})
-                and `meetings`.`status_dict_id` = {$c(Meeting::STATUS_PLANNED)}
-                and `meetings`.`category_dict_id` != {$c(Meeting::CATEGORY_SYSTEM)}
-                and `meetings`.`deleted_at` is null
-                /* Next month */
-                and date > curdate()
-                and date <= date_add(curdate(), interval 1 month)",
-            "planned_meetings_count_next_month"
-        );
     }
 }
